hostfs: fix opening file with overlayfs
If a file on overlayfs using hostfs as a lowerdir is opened, it would
be failed by ENOENT because hostfs resolves a host path by a host root
path stored on the hostfs superblock referred by file->f_path.dentry,
but it is replaced with overlayfs's one starting from the
commit 4bacc9c9234c ("overlayfs: Make f_path always point to the
overlay and f_inode to the underlay").

We replace using file->f_path.dentry directly with calling
"file_dentry" from a commit d101a125954e ("fs: add file_dentry()") to
fix the issue.

Signed-off-by: Toshikuni Fukaya <toshiq2@gmail.com>